refactor(observatory): split heap module
