+++ src/observatory/src/store/heap/mod.rs
@@ -1,0 +1,7 @@
+mod controllers;
+mod env;
+mod openid;
+
+pub use controllers::*;
+pub use env::*;
+pub use openid::*;
